test: read tests.json at runtime
