+++ scripts/run-tests/scan.ts
@@ -1,8 +1,31 @@
 import path from 'path'
 import fs from 'fs'
-import { getAllFiles } from './get-all-files'
 
-const outFile = path.resolve(__dirname, 'tests.json')
+/**
+ * gets all files recursively under the root provided
+ */
+const getAllFiles = (root: string, ignore: string[] = []) => {
+  const allFiles: string[] = []
+  const ls = (cwd: string) => {
+    fs.readdirSync(cwd)
+      .filter((x) => !ignore.includes(x))
+      .forEach((file) => {
+        const filepath = path.resolve(cwd, file)
+        if (fs.lstatSync(filepath).isDirectory()) {
+          ls(path.resolve(cwd, file))
+        } else {
+          allFiles.push(filepath)
+        }
+      })
+  }
+  ls(root)
+  return allFiles
+    .map((path) => path.replace(root + '/', ''))
+    .filter((entry) => entry !== '')
+}
+
+const rootDir = path.resolve(__dirname, '../..')
+const outFile = path.resolve(rootDir, 'tests.json')
 
 /**
  * @param {string} path
@@ -14,8 +37,6 @@ const getTestName = (path: string): string => {
     .match(/displayName: ?["'](.*)["'],/)
   return re ? re[1] : ''
 }
-
-const rootDir = path.resolve(__dirname, '../..')
 
 /**
  * scan for jest configs recursively
